Add files via upload
This is an Retention dataset, after carefully cleaning the dataset, I had to analyze, to find out why are leaving the particular firm. as it is stated that any organization with an retention rate of "84% and above" is a toxic organization. With my analysis we can deduce that this organization is not a toxic environment, with its retention rate at 83.88%
with different reasons for retention. We can see the high earners still staying back, Age was another factor to consider, Education also.